feat: split cart & about code structures to modules
