+++ Online-store/src/pages/about/index.html
@@ -52,7 +52,7 @@
     </div>
   </footer>
 
-  <script src="js/script.js"></script>
+  <script src="js/script.js" type="module"></script>
 </body>
 
 </html>
