+++ Online-store/src/pages/cart/index.html
@@ -108,7 +108,7 @@
     </form>
   </div>
 
-  <script src="js/script.js"></script>
+  <script src="js/script.js" type="module"></script>
 </body>
 
 </html>
